fix: fix bug in determining reviewers to add based on changed files
The previous code had a bug where it would not correctly determine the reviewers to add based on the changed files. The bug was in the line that checked if the folder of a changed file matched any of the folders in the `folders_reviewers` dictionary. The bug was fixed by changing the line to use the `any` function instead of a list comprehension.