Add reviewers per files and folders (#361)
* feat: add support for reviewers defined in OWNERS file for specific files and folders

The code changes include:
- Added a method `get_owners_content` to retrieve the content of the OWNERS file
- Modified the `owners_content` property to call the `get_owners_content` method
- Added new properties `files_reviewers` and `folders_reviewers` to retrieve the reviewers defined for specific files and folders in the OWNERS file
- Modified the `reviewers` property to include reviewers defined in the OWNERS file for specific files and folders
- Added a new method `list_changed_commit_files` to retrieve the list of changed files in the last commit
- Modified the `assign_reviewers` method to assign reviewers based on the OWNERS file for specific files and folders

The OWNERS file has been updated to include reviewers for specific files and folders.

* docs: update OWNERS file example and fix typo in comment

* docs: update OWNERS file example and remove duplicate code block

* fix: fix bug in determining reviewers to add based on changed files

The previous code had a bug where it would not correctly determine the reviewers to add based on the changed files. The bug was in the line that checked if the folder of a changed file matched any of the folders in the `folders_reviewers` dictionary. The bug was fixed by changing the line to use the `any` function instead of a list comprehension.